added confirmation box to create and delete

+++ src/components/Card.js
@@ -4,20 +4,35 @@ import "./Card.scss"
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faPhone, faAt } from '@fortawesome/free-solid-svg-icons';
 import { faGithub, faLinkedin, faFacebook, faInstagram } from "@fortawesome/free-brands-svg-icons"
+import { confirmAlert } from 'react-confirm-alert';
 
 export default function Card(props) {
 
   const handleDelete = e => {
     e.preventDefault();
-    const headers = {
-      headers: { Authorization: `Bearer ${props.currentUser.token}`}
-    };
-    return axios.delete(`http://localhost:8001/api/cards/${props.card.id}`, headers)
-    .then(r => {
-      console.log(r)
-      props.deleteCard(props.card.id)
-    })
-    .catch(e => console.log(e))
+    confirmAlert({
+      title: 'Delete Card',
+      message: 'Are you sure to do this.',
+      buttons: [
+        {
+          label: 'Confirm',
+          onClick: () => {
+            const headers = {
+              headers: { Authorization: `Bearer ${props.currentUser.token}`}
+            };
+            return axios.delete(`http://localhost:8001/api/cards/${props.card.id}`, headers)
+            .then(r => {
+              console.log(r)
+              props.deleteCard(props.card.id)
+            })
+            .catch(e => console.log(e))
+          }
+        },
+        {
+          label: 'Cancel',
+        }
+      ]
+    });
   }
 
   return (
@@ -103,7 +118,9 @@ export default function Card(props) {
             <button>edit card</button>
           </div>
         }
-        <button onClick={handleDelete}>delete card</button>
+          <div>
+            <button onClick={handleDelete}>delete card</button>
+          </div>
       </footer>
     </article>
   )
